Added recurring election year
It's pretty hardcoded right now, thinking down the road or in v2 to go
ahead and move these kind of constants into the country's data files.
Things fall apart when things aren't on a set schedule so this could
even be expanded to allow for a list of incidental years that can be
added to the range when checking. Closes #17 Related to #16