Fixed month view
With multiple holidays on a single day, the array ended up being chunked
out instead of just being an array of the holidays. Got me thinking
about the year view and how it's inconsistent with the other two views.
I'm thinking that the year view should default to the same un-nested
view and then have an optional flag to index on a field (name, or date)
depending on how you want to be using the data.